Switched to using path.join()
As discovered in the recent progress update meeting, Windows machines
have problem with the combination of the glob and require usage that we
had in place for including controllers. This code was rather old and had
been incorporated when scaffolding the skeleton of the original app so
it wouldn't surprise me if there is a better approach, but it was an
easy enough fix to use `path.join()` in place of some of the path
delimiters and string manipulation.

I problem went a bit overboard on the replacements as I know the other
code generally didn't have a problem but I don't think it hurts anything
and is still readable.