WIP: Example Configuration via Serial
Added a lot of code in this commit...

First, started on an initial implementation for reading a configuration
from the serial interface and writing it to the EEPROM for future boots.
This is partially completed in that it can accept some of the serial
input and correctly parse it to prepare to write to the EEPROM. However,
it currently doesn't write the data so after the correct configuration
string is supplied, it tries to connect with the defaul values
(SSID/XXXXXXXX) and it fails.

Also added a lot more informative blinking of the onboard LED so that
the serial monitor isn't the only way to debug issues. This still needs
to be documented but at this point you can read the code to see what it
is doing, as it is pretty straightforward.

Next I will finish writing to the EEPROM when a correct config string is
received and after that I will make sure that the bootup reads from the
EEPROM first. Finally, we will need a way to "reconfigure" a device if
desired, so I will work on that as well.