Added verification of EEPROM data
WIP: BUT, I want the students to see this code so I will merge into the
main branch shortly. The "save" feature needs to be completed so that
the EEPROM data can persist across reboots, and there might need to be
some more error handling / robustness added to the approach. In
addition, I want to create a way to "reset" the config while will
require some changes to the main loop (as it is written, once the config
is saved to the EEPROM it will never accept a new config as it will skip
right to primary processing). So lots more to do, but good examples
throughout.